requirements.txt
numpy==1.24.4
opencv-python==4.12.0.88
packaging==25.0
pandas==2.0.3
pdf2image==1.17.0
pillow==10.4.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
tzdata==2025.3
